fix brand issues

+++ src/main/java/com/nttdata/estore/repositories/CampaignRepository.java
@@ -3,5 +3,5 @@ package com.nttdata.estore.repositories;
 import com.nttdata.estore.entities.Campaign;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CampaignRepository extends CrudRepository<Campaign, Long> {
+public interface CampaignRepository extends CrudRepository<Campaign, Integer> {
 }
